feat : User jwt 적용

+++ backend/src/main/java/com/codestates/user/controller/UserController.java
@@ -6,20 +6,18 @@ import com.codestates.user.dto.UsersAllResponseDtos;
 import com.codestates.user.entity.User;
 import com.codestates.user.mapper.UserMapper;
 import com.codestates.user.service.UserService;
-import io.jsonwebtoken.Claims;
-import io.jsonwebtoken.Jws;
-import io.jsonwebtoken.JwtException;
-import io.jsonwebtoken.Jwts;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
 import javax.validation.constraints.Positive;
 import java.util.List;
+import java.util.Map;
 
 @Slf4j
 @Validated
@@ -72,30 +70,21 @@ public class UserController {
 
     // [회원프로필 수정]
     @PatchMapping("/{user-id}/edit")
-    public ResponseEntity patchUser(@RequestHeader(name="Authorization") String token,
-                                    @PathVariable("user-id") @Positive long userId,
-                                    @Valid @RequestBody UserDto.PatchDto requestBody){
+    public ResponseEntity patchUser(@PathVariable("user-id") @Positive long userId,
+                                    @Valid @RequestBody UserDto.PatchDto requestBody,
+                                    Authentication authentication){
 
-        //Todo: JWT 검증
-        String secretKey = System.getenv("JWT_SECRET_KEY");
-        Jws<Claims> jws;
+        Map<String,Object> principal = (Map) authentication.getPrincipal();
+        long jwtUserId = ((Number) principal.get("userId")).longValue();
 
-        try {
-            jws = Jwts.parserBuilder().setSigningKey(secretKey.getBytes()).build().parseClaimsJws(token);
-        } catch (JwtException e){
-            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        if(jwtUserId != (userId)){
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
         }
-
-        long jwtUserId = Long.parseLong(jws.getBody().getSubject());
-
-        if(jwtUserId != userId){
-            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
-        }
-
-
+        requestBody.setUserId(userId);
         User user = userService.updateUser(userMapper.userPatchDtoToUser(requestBody));
         return new ResponseEntity<>(userMapper.userToUserResponseDto(user),HttpStatus.OK);
     }
+
 
     // [회원조회]
     @GetMapping("/{user-id}")
@@ -118,8 +107,16 @@ public class UserController {
 
     //회원탈퇴
     @DeleteMapping("/{user-id}")
-    public ResponseEntity deleteUser(@PathVariable("user-id") @Positive long userId){
+    public ResponseEntity deleteUser(@PathVariable("user-id") @Positive long userId,
+                                     Authentication authentication){
+
+        Map<String,Object> principal = (Map) authentication.getPrincipal();
+        long jwtUserId = ((Number) principal.get("userId")).longValue();
+
+        if(jwtUserId != userId){
+            return new ResponseEntity(HttpStatus.FORBIDDEN);
+        }
         userService.removeUser(userId);
-        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        return new ResponseEntity<>(HttpStatus.OK);
     }
 }
